requirements.txt
python-daemon==3.0.1
suntime==1.2.5
prettytable==3.8.0
requests-oauthlib==1.3.1
schedule==1.2.0